requirements.txt
docopt
spacy
google-cloud-translate
requests
uuid
